refactor: split oversized modules across frontend and ai services
