+++ src/context/llm/useExecuteCompletion.ts
@@ -32,6 +32,12 @@ import {
   buildServiceRuntimeConfig,
 } from './service-runtime-config';
 import { isSupersededRequestError } from './types';
+import {
+  buildStreamingMessage,
+  extractThinkingText,
+  extractToolCalls,
+  hasRenderableAssistantOutput,
+} from './streaming-message-utils';
 
 const logger = getLogger('useExecuteCompletion');
 
@@ -421,30 +427,16 @@ export function useExecuteCompletion({
                 return prev;
               }
               const next = new Map(prev);
-              const toolCalls: ToolCall[] = content
-                .filter((c) => c.type === 'tool_call')
-                .map((c) => {
-                  const tc = c as MCPToolCallContent;
-                  return {
-                    id: tc.id,
-                    type: 'function',
-                    function: { name: tc.name, arguments: tc.arguments },
-                  };
-                });
-              const thinking = content
-                .filter((c) => c.type === 'thinking')
-                .map((c) => (c as MCPThinkingContent).thinking)
-                .join('\n');
-              next.set(sessionId, {
-                ...streamingMessage,
-                content,
-                tool_calls: toolCalls.length > 0 ? toolCalls : undefined,
-                thinking: thinking || undefined,
-                thinkingSignature,
-                thinkingTime: currentThinkingTime,
-                usage: finalUsage,
-                isStreaming: true,
-              });
+              next.set(
+                sessionId,
+                buildStreamingMessage(
+                  streamingMessage,
+                  content,
+                  thinkingSignature,
+                  currentThinkingTime,
+                  finalUsage,
+                ),
+              );
               return next;
             });
           }
@@ -457,30 +449,16 @@ export function useExecuteCompletion({
             return prev;
           }
           const next = new Map(prev);
-          const toolCalls: ToolCall[] = content
-            .filter((c) => c.type === 'tool_call')
-            .map((c) => {
-              const tc = c as MCPToolCallContent;
-              return {
-                id: tc.id,
-                type: 'function',
-                function: { name: tc.name, arguments: tc.arguments },
-              };
-            });
-          const thinking = content
-            .filter((c) => c.type === 'thinking')
-            .map((c) => (c as MCPThinkingContent).thinking)
-            .join('\n');
-          next.set(sessionId, {
-            ...streamingMessage,
-            content,
-            tool_calls: toolCalls.length > 0 ? toolCalls : undefined,
-            thinking: thinking || undefined,
-            thinkingSignature,
-            thinkingTime: currentThinkingTime,
-            usage: finalUsage,
-            isStreaming: true,
-          });
+          next.set(
+            sessionId,
+            buildStreamingMessage(
+              streamingMessage,
+              content,
+              thinkingSignature,
+              currentThinkingTime,
+              finalUsage,
+            ),
+          );
           return next;
         });
 
@@ -505,21 +483,8 @@ export function useExecuteCompletion({
           }
         }
 
-        const finalToolCalls: ToolCall[] = content
-          .filter((c) => c.type === 'tool_call')
-          .map((c) => {
-            const tc = c as MCPToolCallContent;
-            return {
-              id: tc.id,
-              type: 'function',
-              function: { name: tc.name, arguments: tc.arguments },
-            };
-          });
-
-        const finalThinking = content
-          .filter((c) => c.type === 'thinking')
-          .map((c) => (c as MCPThinkingContent).thinking)
-          .join('\n');
+        const finalToolCalls: ToolCall[] = extractToolCalls(content);
+        const finalThinking = extractThinkingText(content);
 
         const finalMessage: Message = {
           id: responseMessageId,
@@ -529,7 +494,7 @@ export function useExecuteCompletion({
           content,
           createdAt: new Date(),
           tool_calls: finalToolCalls.length > 0 ? finalToolCalls : undefined,
-          thinking: finalThinking || undefined,
+          thinking: finalThinking,
           thinkingSignature,
           thinkingTime: thinkingStartTime
             ? (performance.now() - thinkingStartTime) / 1000
@@ -553,13 +518,7 @@ export function useExecuteCompletion({
             : undefined,
         });
 
-        const hasContent =
-          (finalMessage.content &&
-            finalMessage.content.some((c) =>
-              c.type === 'text' ? !!(c as MCPTextContent).text?.trim() : true,
-            )) ||
-          (finalMessage.tool_calls && finalMessage.tool_calls.length > 0) ||
-          !!finalMessage.thinking;
+        const hasContent = hasRenderableAssistantOutput(finalMessage);
 
         const hasUsage =
           finalMessage.usage && finalMessage.usage.completionTokens > 0;
