+++ src/features/agent/hooks/useAgentDraftChat.ts
@@ -13,13 +13,7 @@ import type { AssistantDto } from '@/lib/backend/assistants';
 import { parseAssistant } from '@/models/validation';
 import { useSettings } from '@/context/SettingsContext';
 import { enforceRuntimeBuiltinAliases } from '@/lib/assistant/runtime-builtins';
-import {
-  workspaceWriteFile,
-  getWorkspaceDir,
-  checkDroppedPathType,
-} from '@/lib/backend';
-import { generateWorkspacePath } from '@/lib/workspace-sync-service';
-import { workspacePathToFileUrl } from '@/lib/file-url';
+import { checkDroppedPathType } from '@/lib/backend';
 import { getMimeTypeFromFilename } from '@/lib/mime-utils';
 import {
   useDnDContext,
@@ -27,19 +21,12 @@ import {
   type DragAndDropPayload,
 } from '@/context/DnDContext';
 import { useRustBackend } from '@/hooks/use-rust-backend';
-import { saveAgentFile } from '@/features/agent/api/agent-backend';
 import { useInputToken } from './useInputToken';
 import { useScopedSkills } from './useScopedSkills';
 import type { AgentEventPayload } from '@/context/AgentSessionContext';
+import { prepareDraftAttachments } from '../lib/draft-attachments';
 
 const logger = getLogger('useAgentDraftChat');
-
-// File extensions treated as plain text for Attachments indexing
-const TEXT_EXTENSIONS_DRAFT =
-  /\.(txt|md|markdown|json|jsonc|json5|yaml|yml|toml|js|jsx|ts|tsx|mjs|cjs|py|rb|rs|go|java|c|cpp|h|hpp|css|scss|less|html|htm|svg|sh|bash|zsh|fish|ps1|sql|graphql|csv|log|xml|proto)$/i;
-
-// Binary file extensions that the backend document parser can handle via file:// URL
-const BINARY_INDEXABLE_EXTENSIONS_DRAFT = /\.(pdf|docx|xlsx)$/i;
 
 export interface BuiltinServerInfo {
   name: string; // This is the ID
@@ -380,77 +367,16 @@ export function useAgentDraftChat() {
           },
         });
 
-        const attachments: AttachmentReference[] = [];
-        const workspaceDir = await getWorkspaceDir(newSessionId);
-        for (const file of pendingFiles) {
-          try {
-            const workspacePath = generateWorkspacePath(file.name);
-            const arrayBuffer = await file.arrayBuffer();
-            const bytes = Array.from(new Uint8Array(arrayBuffer));
-            await workspaceWriteFile(workspacePath, bytes, newSessionId);
-            let lineCount = 0;
-            const isText =
-              /^text\/|\/(json|xml|javascript|typescript)/.test(file.type) ||
-              TEXT_EXTENSIONS_DRAFT.test(file.name);
-            if (isText) {
-              try {
-                const text = await file.text();
-                lineCount = text.split('\n').length;
-              } catch {
-                // non-critical, leave as 0
-              }
-            }
-            const actualMimeType =
-              file.type && file.type !== 'application/octet-stream'
-                ? file.type
-                : getMimeType(file.name);
-
-            const isInlineType =
-              actualMimeType.startsWith('image/') ||
-              actualMimeType.startsWith('audio/');
-
-            if (isInlineType) {
-              const inlineType = actualMimeType.startsWith('image/')
-                ? ('image' as const)
-                : ('audio' as const);
-              const fileUrl = workspacePathToFileUrl(
-                workspaceDir,
-                workspacePath,
-              );
-              attachments.push({
-                sessionId: newSessionId,
-                filename: file.name,
-                mimeType: actualMimeType,
-                size: file.size,
-                lineCount: 0,
-                preview: file.name,
-                uploadedAt: now.toISOString(),
-                status: 'inline',
-                workspacePath,
-                inlineContent: {
-                  type: inlineType,
-                  uri: fileUrl,
-                  mimeType: actualMimeType,
-                },
-              });
-            } else {
-              attachments.push({
-                sessionId: newSessionId,
-                filename: file.name,
-                mimeType: actualMimeType,
-                size: file.size,
-                lineCount,
-                preview: file.name,
-                uploadedAt: now.toISOString(),
-                status: 'workspace-only',
-                workspacePath,
-              });
-            }
-          } catch (err) {
-            logger.error('Failed to write pre-session attachment', err);
-            toast.error(t('agent.draft.failedToAttach', { file: file.name }));
-          }
-        }
+        const attachments: AttachmentReference[] =
+          await prepareDraftAttachments({
+            files: pendingFiles,
+            sessionId: newSessionId,
+            now,
+            getMimeType,
+            onAttachmentError: (file) => {
+              toast.error(t('agent.draft.failedToAttach', { file: file.name }));
+            },
+          });
 
         const initialMessage: Message = {
           id: createId(),
@@ -468,100 +394,6 @@ export function useAgentDraftChat() {
           createdAt: now.getTime(),
           updatedAt: now.getTime(),
         };
-
-        let workspaceDirCache: string | null = null;
-        const getWorkspaceDirCached = async (): Promise<string> => {
-          if (workspaceDirCache === null) {
-            workspaceDirCache = await getWorkspaceDir(newSessionId);
-          }
-          return workspaceDirCache!;
-        };
-
-        for (let i = 0; i < pendingFiles.length; i++) {
-          const file = pendingFiles[i];
-          const isTextFile =
-            TEXT_EXTENSIONS_DRAFT.test(file.name) || /^text\//.test(file.type);
-          const isBinaryIndexable = BINARY_INDEXABLE_EXTENSIONS_DRAFT.test(
-            file.name,
-          );
-
-          if (isTextFile) {
-            try {
-              const content = await file.text();
-              const result = await saveAgentFile(newSessionId, file.name, {
-                content,
-                metadata: {
-                  mimeType: file.type || 'text/plain',
-                  size: file.size,
-                  uploadedAt: now.toISOString(),
-                  filename: file.name,
-                },
-              });
-
-              // Runtime validation for contentId existence
-              if (
-                result &&
-                typeof result === 'object' &&
-                'contentId' in result
-              ) {
-                attachments[i] = {
-                  ...attachments[i],
-                  status: 'committed',
-                  contentId: result.contentId,
-                  lineCount: result.lineCount ?? attachments[i].lineCount,
-                };
-              }
-            } catch (commitErr) {
-              logger.warn(
-                'Failed to commit file to Attachments store, keeping workspace-only',
-                { filename: file.name, error: commitErr },
-              );
-            }
-          } else if (isBinaryIndexable) {
-            try {
-              const workspaceDir = await getWorkspaceDirCached();
-              const workspacePath = attachments[i].workspacePath;
-              if (!workspacePath) {
-                logger.warn(
-                  'Binary file missing workspacePath, skipping index',
-                  { filename: file.name },
-                );
-                continue;
-              }
-              const normalizedDir = workspaceDir.replace(/\\/g, '/');
-              const normalizedRelative = workspacePath.replace(/\\/g, '/');
-              const fileUrl = `file:///${normalizedDir.replace(/^\//, '')}/${normalizedRelative}`;
-              const result = await saveAgentFile(newSessionId, file.name, {
-                fileUrl,
-                metadata: {
-                  mimeType: file.type || 'application/octet-stream',
-                  size: file.size,
-                  uploadedAt: now.toISOString(),
-                  filename: file.name,
-                },
-              });
-
-              // Runtime validation for contentId existence
-              if (
-                result &&
-                typeof result === 'object' &&
-                'contentId' in result
-              ) {
-                attachments[i] = {
-                  ...attachments[i],
-                  status: 'committed',
-                  contentId: result.contentId,
-                  lineCount: result.lineCount ?? attachments[i].lineCount,
-                };
-              }
-            } catch (commitErr) {
-              logger.warn(
-                'Failed to commit binary file to Attachments store, keeping workspace-only',
-                { filename: file.name, error: commitErr },
-              );
-            }
-          }
-        }
 
         const finalRustMessage = {
           ...rustMessage,
