+++ src/features/knowledge/components/DeleteKnowledgeDialog.tsx
@@ -1,0 +1,62 @@
+import { Loader2 } from 'lucide-react';
+import { useTranslation } from 'react-i18next';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog';
+
+interface DeleteKnowledgeDialogProps {
+  isDeleting: boolean;
+  onConfirm: () => void;
+  onOpenChange: (open: boolean) => void;
+  open: boolean;
+}
+
+export function DeleteKnowledgeDialog({
+  isDeleting,
+  onConfirm,
+  onOpenChange,
+  open,
+}: DeleteKnowledgeDialogProps) {
+  const { t } = useTranslation('common');
+
+  return (
+    <AlertDialog open={open} onOpenChange={onOpenChange}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>
+            {t('knowledge.confirmDeleteTitle', 'Delete knowledge entry')}
+          </AlertDialogTitle>
+          <AlertDialogDescription>
+            {t(
+              'knowledge.confirmDelete',
+              'Delete this knowledge entry and clean up orphaned graph data?',
+            )}
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={isDeleting}>
+            {t('knowledge.cancel', 'Cancel')}
+          </AlertDialogCancel>
+          <AlertDialogAction
+            onClick={(event) => {
+              event.preventDefault();
+              onConfirm();
+            }}
+            disabled={isDeleting}
+            className="gap-2"
+          >
+            {isDeleting ? <Loader2 className="h-4 w-4 animate-spin" /> : null}
+            {t('knowledge.delete', 'Delete')}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
