+++ src/models/service-context.ts
@@ -1,0 +1,4 @@
+export interface ServiceContext {
+  contextPrompt: string;
+  structuredState?: Record<string, unknown>;
+}
